modules/java/static/html/core/src/main/java/org/metawidget/statically/html/widgetprocessor/IdProcessor.java
// Metawidget
//
// This file is dual licensed under both the LGPL
// (http://www.gnu.org/licenses/lgpl-2.1.html) and the EPL
// (http://www.eclipse.org/org/documents/epl-v10.php). As a
// recipient of Metawidget, you may choose to receive it under either
// the LGPL or the EPL.
//
// Commercial licenses are also available. See http://metawidget.org
// for details.

package org.metawidget.statically.html.widgetprocessor;

import static org.metawidget.inspector.InspectionResultConstants.*;

import java.util.Map;

import org.metawidget.statically.StaticXmlWidget;
import org.metawidget.statically.html.StaticHtmlMetawidget;
import org.metawidget.statically.html.widgetbuilder.IdHolder;
import org.metawidget.widgetprocessor.iface.WidgetProcessor;

/**
 * WidgetProcessor that adds an 'id' attribute to a StaticXmlWidget.
 *
 * @author <a href="http://kennardconsulting.com">Richard Kennard</a>
 */

public class IdProcessor
	implements WidgetProcessor<StaticXmlWidget, StaticHtmlMetawidget> {

	//
	// Public methods
	//

	public StaticXmlWidget processWidget( StaticXmlWidget widget, String elementName, Map<String, String> attributes, StaticHtmlMetawidget metawidget ) {

		if ( widget instanceof IdHolder ) {

			String id = attributes.get( NAME );

			if ( id == null ) {
				id = metawidget.getId();
			} else if ( metawidget.getId() != null ) {
				id = metawidget.getId() + '-' + id;
			}

			((IdHolder) widget).setId( id );
		}

		return widget;
	}
}
